Move components with GraphQL queries into componnets
