+++ src/services/useDebounce.js
@@ -1,5 +1,3 @@
-// @flow
-
 import * as React from "react"
 
 const useDebounce = (value, delay) => {
